+++ website/src/components/McapRecordingDemo/hooks/index.ts
@@ -1,0 +1,6 @@
+export { useRecordClick } from "./useRecordClick";
+export { useDownloadClick } from "./useDownloadClick";
+export { useMouseEventRecording } from "./useMouseEventRecording";
+export { useDeviceOrientationRecording } from "./useDeviceOrientationRecording";
+export { useVideoRecording } from "./useVideoRecording";
+export { useAudioRecording } from "./useAudioRecording";
